Routers, cleaningcode, new files etc...
